parser: Move the `grammar` module one level up.
